chore: move existing task implementation to redis
